+++ jeap-cli/src/main/java/ch/admin/bit/jeap/cli/backfill/PasBackfillException.java
@@ -1,0 +1,8 @@
+package ch.admin.bit.jeap.cli.backfill;
+
+public class PasBackfillException extends RuntimeException {
+
+    public PasBackfillException(String message) {
+        super(message);
+    }
+}