+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.3]
+
+### Added
+
+- Added `pas-backfill send` and `pas-backfill report` commands for submitting PAS backfill YAML jobs and reading reports.
+- Added `pas-backfill-db` starter library
+- Added documentation for the PAS backfill job YAML structure.
+
+### Changed
+
+- Updated Readme and Docs
+
 ## [1.6.2] - 2026-06-17
 
 ### Changed